+++ Tests/Tests/Entitas/Fixtures/EntityEmittingSubSystem.cs
@@ -4,23 +4,21 @@ using Entitas;
 public class EntityEmittingSubSystem : IReactiveSystem {
     public int didExecute { get { return _didExecute; } }
 
-    public Entity[] entites { get { return _entites; } }
+    public Entity[] entities { get { return _entities; } }
 
     readonly Pool _pool;
     int _didExecute;
-    Entity[] _entites;
+    Entity[] _entities;
 
     public EntityEmittingSubSystem(Pool pool) {
         _pool = pool;
     }
 
-    public IMatcher trigger { get { return Matcher.AllOf(new [] { CID.ComponentA }); } }
-
-    public GroupEventType eventType { get { return GroupEventType.OnEntityAdded; } }
+    public TriggerOnEvent trigger { get { return Matcher.AllOf(new [] { CID.ComponentA }).OnEntityAdded(); } }
 
     public void Execute(List<Entity> entities) {
         _pool.CreateEntity().AddComponent(CID.ComponentA, new ComponentA());
-        _entites = entities.ToArray();
+        _entities = entities.ToArray();
         _didExecute++;
     }
 }