+++ Tests/Tests/Entitas/Fixtures/InitializeSystemSpy.cs
@@ -1,0 +1,12 @@
+﻿using Entitas;
+
+public class InitializeSystemSpy : IInitializeSystem {
+    public bool initialized { get { return _initialized; } }
+
+    bool _initialized;
+
+    public void Initialize() {
+        _initialized = true;
+    }
+}
+