+++ Tests/Tests/Entitas/describe_GroupObserver.cs
@@ -39,7 +39,7 @@ class describe_GroupObserver : nspec {
                 entities.should_contain(e);
             };
 
-            it["collects entites only once"] = () => {
+            it["collects entities only once"] = () => {
                 var e = pool.CreateEntity();
                 e.AddComponentA();
                 e.RemoveComponentA();
@@ -84,12 +84,33 @@ class describe_GroupObserver : nspec {
                 entities.should_contain(e2);
             };
 
-            it["clears collected entites"] = () => {
+            it["clears collected entities"] = () => {
                 var e = pool.CreateEntity();
                 e.AddComponentA();
 
                 observer.ClearCollectedEntities();
                 observer.collectedEntities.should_be_empty();
+            };
+            
+            context["reference counting"] = () => {
+                it["keeps reference count of an entity at one even after destroy"] = () => {
+                    var e = pool.CreateEntity();
+                    e.AddComponentA();
+                    e.OnEntityReleased += entity => this.Fail();
+                    pool.DestroyEntity(e);
+                };
+                
+                it["counts entity reference down when clearing"] = () => {
+                    Entity eventEntity = null;
+                    var e = pool.CreateEntity();
+                    e.OnEntityReleased += entity => {
+                        eventEntity = entity;
+                    };
+                    pool.DestroyEntity(e);
+                    observer.ClearCollectedEntities();
+                    eventEntity.should_not_be_null();
+                    eventEntity.should_be_same(e);
+                };
             };
         };