+++ Tests/Tests/Entitas/describe_ReactiveSystem.cs
@@ -46,6 +46,17 @@ class describe_ReactiveSystem : nspec {
                 subSystem.entities.should_contain(e);
             };
 
+            it["collects changed entities in execute"] = () => {
+                subSystem.replaceComponentAOnExecute = true;
+                var e = pool.CreateEntity();
+                e.AddComponentA();
+                e.AddComponentB();
+                reactiveSystem.Execute();
+                reactiveSystem.Execute();
+
+                subSystem.didExecute.should_be(2);
+            };
+
             it["doesn't execute when triggeringMatcher doesn't match"] = () => {
                 var e = pool.CreateEntity();
                 e.AddComponentA();
@@ -120,6 +131,28 @@ class describe_ReactiveSystem : nspec {
                 subSystem.didExecute.should_be(0);
                 subSystem.entities.should_be_null();
             };
+
+            it["retains entities until execute completed"] = () => {
+                var didExecute = false;
+                subSystem.executeBlock = entities => {
+                    var providedEntity = entities[0];
+                    var newEntitty = pool.CreateEntity();
+                    providedEntity.should_not_be_null();
+                    providedEntity.should_not_be_same(newEntitty);
+                    didExecute = true;
+                };
+                
+                var e = pool.CreateEntity();
+                e.AddComponentA();
+                e.AddComponentB();
+                pool.DestroyEntity(e);
+                reactiveSystem.Execute();
+                didExecute.should_be_true();
+
+                var reusedEntity = pool.CreateEntity();
+                reusedEntity.should_not_be_null();
+                reusedEntity.should_be_same(e);
+            };
         };
 
         context["OnEntityAddedOrRemoved"] = () => {
@@ -157,23 +190,19 @@ class describe_ReactiveSystem : nspec {
                 e.AddComponentA();
                 e.AddComponentB();
                 reactiveSystem.Execute();
-                sys.entites.Length.should_be(1);
+                sys.entities.Length.should_be(1);
                 reactiveSystem.Execute();
-                sys.entites.Length.should_be(1);
+                sys.entities.Length.should_be(1);
             };
         };
 
         context["MultiReactiveSystem"] = () => {
             before = () => {
-                var matchers = new IMatcher[] {
-                    Matcher.AllOf(new [] { CID.ComponentA }),
-                    Matcher.AllOf(new [] { CID.ComponentB })
+                var triggers = new [] {
+                    Matcher.AllOf(new [] { CID.ComponentA }).OnEntityAdded(),
+                    Matcher.AllOf(new [] { CID.ComponentB }).OnEntityRemoved()
                 };
-                var eventTypes = new [] {
-                    GroupEventType.OnEntityAdded,
-                    GroupEventType.OnEntityRemoved,
-                };
-                multiSubSystem = new MultiReactiveSubSystemSpy(matchers, eventTypes);
+                multiSubSystem = new MultiReactiveSubSystemSpy(triggers);
                 reactiveSystem = new ReactiveSystem(pool, multiSubSystem);
             };
             
@@ -199,7 +228,7 @@ class describe_ReactiveSystem : nspec {
 
         context["ensure components matcher"] = () => {
 
-            it["only passes in entites matching required matcher"] = () => {
+            it["only passes in entities matching required matcher"] = () => {
                 var ensureSubSystem = new ReactiveEnsureSubSystemSpy(
                     allOfAB(),
                     GroupEventType.OnEntityAdded,
@@ -225,23 +254,44 @@ class describe_ReactiveSystem : nspec {
                 ensureSubSystem.entities.should_contain(eABC);
             };
 
-            it["only passes in entites matching required matcher (multi reactive)"] = () => {
-                var matchers = new IMatcher[] {
-                    Matcher.AllOf(new [] { CID.ComponentA }),
-                    Matcher.AllOf(new [] { CID.ComponentB })
-                };
-                var eventTypes = new [] {
+            it["retains entities until execute completed"] = () => {
+                var ensureSubSystem = new ReactiveEnsureSubSystemSpy(
+                    allOfAB(),
                     GroupEventType.OnEntityAdded,
-                    GroupEventType.OnEntityRemoved,
-                };
+                    Matcher.AllOf(new [] {
+                        CID.ComponentA,
+                        CID.ComponentB,
+                        CID.ComponentC
+                    })
+                );
+                reactiveSystem = new ReactiveSystem(pool, ensureSubSystem);
 
+                var eABC = pool.CreateEntity();
+
+                eABC.AddComponentA();
+                eABC.AddComponentB();
+                eABC.AddComponentC();                
+                var refCountBefore = eABC.GetRefCount();
+                refCountBefore.should_be(3); // referd by pool, group and group observer
+
+                reactiveSystem.Execute();
+                var refCountAfter = eABC.GetRefCount();
+               
+                refCountAfter.should_be(2); // refered by pool and group
+            };
+
+            it["only passes in entities matching required matcher (multi reactive)"] = () => {
+                var triggers = new [] {
+                    Matcher.AllOf(new [] { CID.ComponentA }).OnEntityAdded(),
+                    Matcher.AllOf(new [] { CID.ComponentB }).OnEntityRemoved()
+                };
                 var ensure = Matcher.AllOf(new [] {
                     CID.ComponentA,
                     CID.ComponentB,
                     CID.ComponentC
                 });
 
-                var ensureSubSystem = new MultiReactiveEnsureSubSystemSpy(matchers, eventTypes, ensure);
+                var ensureSubSystem = new MultiReactiveEnsureSubSystemSpy(triggers, ensure);
                 reactiveSystem = new ReactiveSystem(pool, ensureSubSystem);
 
                 var eAB = pool.CreateEntity();
@@ -257,10 +307,35 @@ class describe_ReactiveSystem : nspec {
                 ensureSubSystem.entities.Length.should_be(1);
                 ensureSubSystem.entities.should_contain(eABC);
             };
+
+            it["doesn't call execute when no entities left after filtering"] = () => {
+                var ensureSubSystem = new ReactiveEnsureSubSystemSpy(
+                    allOfAB(),
+                    GroupEventType.OnEntityAdded,
+                    Matcher.AllOf(new [] {
+                        CID.ComponentA,
+                        CID.ComponentB,
+                        CID.ComponentC,
+                        CID.ComponentD
+                    })
+                );
+                reactiveSystem = new ReactiveSystem(pool, ensureSubSystem);
+
+                var eAB = pool.CreateEntity();
+                eAB.AddComponentA();
+                eAB.AddComponentB();
+                var eABC = pool.CreateEntity();
+                eABC.AddComponentA();
+                eABC.AddComponentB();
+                eABC.AddComponentC();
+                reactiveSystem.Execute();
+
+                ensureSubSystem.didExecute.should_be(0);
+            };
         };
 
         context["exlude components"] = () => {
-            it["only passes in entites not matching matcher"] = () => {
+            it["only passes in entities not matching matcher"] = () => {
                 var excludeSubSystem = new ReactiveExcludeSubSystemSpy(
                     allOfAB(),
                     GroupEventType.OnEntityAdded,
@@ -282,21 +357,51 @@ class describe_ReactiveSystem : nspec {
                 excludeSubSystem.entities.should_contain(eAB);
             };
 
-            it["only passes in entites not matching required matcher (multi reactive)"] = () => {
-                var matchers = new IMatcher[] {
-                    Matcher.AllOf(new [] { CID.ComponentA }),
-                    Matcher.AllOf(new [] { CID.ComponentB })
-                };
-                var eventTypes = new [] {
-                    GroupEventType.OnEntityAdded,
+            it["retains entities until execute completed"] = () => {
+                var excludeSubSystem = new ReactiveExcludeSubSystemSpy(
+                    allOfAB(),
                     GroupEventType.OnEntityRemoved,
+                    Matcher.AllOf(new [] { CID.ComponentC })
+                );
+                var didExecute = false;
+                excludeSubSystem.executeBlock = entities => {
+                    entities.Count.should_be(1);
+                    var providedEntity = entities[0];
+                    var newEntitty = pool.CreateEntity();
+                    providedEntity.should_not_be_null();
+                    providedEntity.should_not_be_same(newEntitty);
+                    didExecute = true;
                 };
+                reactiveSystem = new ReactiveSystem(pool, excludeSubSystem);
+                
 
+                var e1 = pool.CreateEntity();
+                e1.AddComponentA();
+                e1.AddComponentB();
+                pool.DestroyEntity(e1);
+                var e2 = pool.CreateEntity();
+                e2.AddComponentA();
+                e2.AddComponentB();
+                e2.AddComponentC();
+                e2.RemoveComponentB();
+                reactiveSystem.Execute();
+                didExecute.should_be_true();
+
+                var reusedEntity = pool.CreateEntity();
+                reusedEntity.should_not_be_null();
+                reusedEntity.should_be_same(e1);
+            };
+
+            it["only passes in entities not matching required matcher (multi reactive)"] = () => {
+                var triggers = new [] {
+                    Matcher.AllOf(new [] { CID.ComponentA }).OnEntityAdded(),
+                    Matcher.AllOf(new [] { CID.ComponentB }).OnEntityRemoved()
+                };
                 var exclude = Matcher.AllOf(new [] {
                     CID.ComponentC
                 });
 
-                var excludeSubSystem = new MultiReactiveExcludeSubSystemSpy(matchers, eventTypes, exclude);
+                var excludeSubSystem = new MultiReactiveExcludeSubSystemSpy(triggers, exclude);
                 reactiveSystem = new ReactiveSystem(pool, excludeSubSystem);
 
                 var eAB = pool.CreateEntity();
@@ -315,7 +420,7 @@ class describe_ReactiveSystem : nspec {
         };
 
         context["ensure / exlude components mix"] = () => {
-            it["only passes in entites"] = () => {
+            it["only passes in entities"] = () => {
                 var ensureExcludeSystem = new ReactiveEnsureExcludeSubSystemSpy(
                     allOfAB(),
                     GroupEventType.OnEntityAdded,
@@ -340,6 +445,28 @@ class describe_ReactiveSystem : nspec {
                 ensureExcludeSystem.entities.Length.should_be(1);
                 ensureExcludeSystem.entities.should_contain(eAB);
             };
+
+            it["retains entities until execute completed"] = () => {
+                var ensureExcludeSystem = new ReactiveEnsureExcludeSubSystemSpy(
+                    allOfAB(),
+                    GroupEventType.OnEntityAdded,
+                    Matcher.AllOf(new [] { CID.ComponentA, CID.ComponentB }),
+                    Matcher.AllOf(new [] { CID.ComponentC })
+                );
+                reactiveSystem = new ReactiveSystem(pool, ensureExcludeSystem);
+
+                var eAB = pool.CreateEntity();
+                eAB.AddComponentA();
+                eAB.AddComponentB();
+                var refCountBefore = eAB.GetRefCount();
+                refCountBefore.should_be(3); // referd by pool, group and group observer
+                
+                reactiveSystem.Execute();
+
+                var refCountAfter = eAB.GetRefCount();
+               
+                refCountAfter.should_be(2); // refered by pool and group
+            };
         };
     }
 
@@ -363,3 +490,9 @@ class describe_ReactiveSystem : nspec {
     }
 }
 
+public static class EntityExtgensions {
+    public static int GetRefCount(this Entity e) {
+        return (int)e.GetType().GetField("_refCount", 
+            System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance).GetValue(e);
+    }
+}